Rewrote method to accept empty lists as input
The previous build-method was limited in the way that it only accepted
null-values as a valid input if the user wanted to retrieve all
information for a specific input. Now it is possibly to also specify an
empty list as input, which is equivalent.

The unit test for this is located in QueryBuilderTest and the
integration test is located in AbstractClientIT